clang depfile support
